patching up some tests and steamworks dll issues as well as fixing up soem regressions and making the async event pump actually work

+++ Source/Nomad.Core/Numerics/PackedVector2U16.cs
@@ -71,8 +71,8 @@ namespace Nomad.Core.Numerics
         private static float DequantizeSignedNormalized8(byte value)
         {
             // [0, 255] -> [0, 1] -> [-1, 1]
-            float normalized = value / byte.MaxValue;
-            return (normalized * 2.0f) - 1.0f;
+            double normalized = value / (double)byte.MaxValue;
+            return (float)((normalized * 2.0d) - 1.0d);
         }
     }
 }
